Add files via upload
The dataset given in the question is c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22). I named my dataset as "get" where I called the values of dataset. Then as per the directions given in the question, I assigned a function called myMean where I inserted the logic of mean (sum of dataset divided by the number of objects). Now I performed this function on the given dataset and obtained the output.