Delete assignment2.R
The given dataset in the question is c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22). I have given a variable name as "get" and called all the values in the dataset into my variable. Now, as per the directions in the question, I have declared a function called myMean with the logic of mean(sum of objects divided by number of objects). Since the logic is inserted, the function "myMean" works accordingly and gave me the output with the operation "myMean(get)". The code works successfully.