result
The given dataset in the question is  c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22). I have assigned a variable named "get" and loaded all the values in the dataset to this variable. Later , as per the directions given in the question , I created a function named "myMean" which executes the mean for the dataset with the provided logic in the question. Now the output got successfully printed the output of mean 19.25,since the function is properly created with the operation given in the question.